Static Libraries and ZERO Warnings Let Hazel be Static library and Not support DLL now Handle warnings C4244, C4251, C4819

+++ Hazel/src/Platform/Windows/WindowsWindow.cpp
@@ -134,7 +134,7 @@ namespace Hazel
 		{
 			WindowData& data = *(WindowData*)glfwGetWindowUserPointer(window);
 
-			MouseScrolledEvent event(xoffset, yoffset);
+			MouseScrolledEvent event((float)xoffset, (float)yoffset);
 			data.EventCallback(event);
 		});
